requirements.txt
lightning
numpy<2
pandas
python-dotenv
pytorch-forecasting
scikit-learn
supabase
torch